Deploy

+++ migrations/20180927101219_Users.js
@@ -1,16 +1,16 @@
 
-exports.up = function(knex, Promise) {
-    return knex.schema.createTable('Users', function (table){
+exports.up = function (knex, Promise) {
+    return knex.schema.createTable('Users', function (table) {
         table.increments().primary()
         table.string('username')
         table.string('password')
         table.string('email')
         table.integer('coins')
-        table.integer('logoutTime')
+        table.bigInteger('logoutTime')
     })
-  
+
 }
 
-exports.down = function(knex, Promise) {
+exports.down = function (knex, Promise) {
     return knex.schema.dropTable('Users')
 };
